Fix problem with matplotlib import in postprocess.py
During import of matplotlib.pyplot in module/src/postprocess.py a KeyError was thrown because the environment variable "WINDIR" is missing. This was caused by the fact that only a small selection of the environment variables is available in when the interpreter for CascadeToxswa is run. This adaptation assures that the WINDIR and USERPROFILE environment variables are available.